% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_methods.R
\name{as_df_list}
\alias{as_df_list}
\title{Create a df_list list}
\usage{
as_df_list(l)
}
\arguments{
\item{l}{a list of data frames}
}
\value{
a list of data frames with class df_list
}
\description{
Create a df_list list
}
\examples{
as_df_list(list(a = mtcars, b = mtcars))
}
