% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_methods.R
\name{filter.df_list}
\alias{filter.df_list}
\title{Filter data frames in a list}
\usage{
\method{filter}{df_list}(.data, ...)
}
\arguments{
\item{.data}{A data frame, data frame extension (e.g. a tibble), or a
lazy data frame (e.g. from dbplyr or dtplyr). See \emph{Methods}, below, for
more details.}

\item{...}{<\code{\link[dplyr:dplyr_data_masking]{data-masking}}> Expressions that return a
logical value, and are defined in terms of the variables in \code{.data}.
If multiple expressions are included, they are combined with the \code{&} operator.
Only rows for which all conditions evaluate to \code{TRUE} are kept.}
}
\value{
A list of filtered data frames
}
\description{
See \link[dplyr]{filter}
}
\examples{
library(dplyr)
filter(
  as_df_list(list(a = mtcars, b = mtcars)),
  cyl == 4
)
}
