% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_methods.R
\name{mutate.df_list}
\alias{mutate.df_list}
\alias{transmute.df_list}
\title{Modify or create columns for data frames in a list}
\usage{
\method{mutate}{df_list}(.data, ...)

\method{transmute}{df_list}(.data, ...)
}
\arguments{
\item{.data}{A data frame, data frame extension (e.g. a tibble), or a
lazy data frame (e.g. from dbplyr or dtplyr). See \emph{Methods}, below, for
more details.}

\item{...}{<\code{\link[dplyr:dplyr_data_masking]{data-masking}}> Name-value pairs.
The name gives the name of the column in the output.

The value can be:
\itemize{
\item A vector of length 1, which will be recycled to the correct length.
\item A vector the same length as the current group (or the whole data frame
if ungrouped).
\item \code{NULL}, to remove the column.
\item A data frame or tibble, to create multiple columns in the output.
}}
}
\value{
A list of mutated data frames
}
\description{
See \link[dplyr]{mutate}
}
\examples{
library(dplyr)
mutate(
  as_df_list(list(a = mtcars, b = mtcars)),
  lp100km = 235.214538 / mpg
)
library(dplyr)
transmute(
  as_df_list(list(a = mtcars, b = mtcars)),
  cyl, mpg, lp100km = 235.214538 / mpg
)
}
