% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_methods.R
\name{rename.df_list}
\alias{rename.df_list}
\title{Rename columns of data frames in a list}
\usage{
\method{rename}{df_list}(.data, ...)
}
\arguments{
\item{.data}{A data frame, data frame extension (e.g. a tibble), or a
lazy data frame (e.g. from dbplyr or dtplyr). See \emph{Methods}, below, for
more details.}

\item{...}{For \code{rename()}: <\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> Use
\code{new_name = old_name} to rename selected variables.

For \code{rename_with()}: additional arguments passed onto \code{.fn}.}
}
\value{
A list of data frames with renamed columns
}
\description{
See \link[dplyr]{rename}
}
\examples{
library(dplyr)
rename(
  as_df_list(list(a = mtcars, b = mtcars)),
  num_cyl = cyl, horsepower = hp
)
}
