% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_methods.R
\name{select.df_list}
\alias{select.df_list}
\title{Select columns from data frames in a list}
\usage{
\method{select}{df_list}(.data, ...)
}
\arguments{
\item{.data}{A data frame, data frame extension (e.g. a tibble), or a
lazy data frame (e.g. from dbplyr or dtplyr). See \emph{Methods}, below, for
more details.}

\item{...}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> One or more unquoted
expressions separated by commas. Variable names can be used as if they
were positions in the data frame, so expressions like \code{x:y} can
be used to select a range of variables.}
}
\value{
A list of data frames with the selected columns
}
\description{
See \link[dplyr]{select}
}
\examples{
library(dplyr)
select(
  as_df_list(list(a = mtcars, b = mtcars)),
  cyl, mpg
)
}
